# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Women's Beauty Salon Management SaaS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List all clients for the current salon
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create a new client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClientBody"
      responses:
        "201":
          description: Client created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get a client by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "404":
          description: Not found
    patch:
      operationId: updateClient
      tags: [clients]
      summary: Update a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClientBody"
      responses:
        "200":
          description: Updated client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Delete a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /clients/{id}/appointments:
    get:
      operationId: getClientAppointments
      tags: [clients]
      summary: Get appointment history for a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AppointmentDetail"

  # ── Services ───────────────────────────────────────────────────────────────
...
  /appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      summary: List appointments with optional date filter
      parameters:
        - name: date
          in: query
          schema:
            type: string
            description: ISO date (YYYY-MM-DD)
        - name: status
          in: query
          schema:
            type: string
            enum: [scheduled, confirmed, completed, cancelled]
        - name: employeeId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AppointmentDetail"
    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Create an appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAppointmentBody"
      responses:
        "201":
          description: Appointment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentDetail"
  /appointments/{id}:
    get:
      operationId: getAppointment
      tags: [appointments]
      summary: Get appointment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentDetail"
    patch:
      operationId: updateAppointment
      tags: [appointments]
      summary: Update appointment (status, reschedule, etc.)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAppointmentBody"
      responses:
        "200":
          description: Updated appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentDetail"
    delete:
      operationId: deleteAppointment
      tags: [appointments]
      summary: Delete/cancel an appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Payments ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Client:
      type: object
      properties:
        id:
          type: integer
        salonId:
          type: integer
        name:
          type: string
        phone:
          type: ["string", "null"]
        whatsapp:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, salonId, name, createdAt]
    CreateClientBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        whatsapp:
          type: string
        email:
          type: string
        notes:
          type: string
      required: [name]
    UpdateClientBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        whatsapp:
          type: string
        email:
          type: string
        notes:
          type: string

    # ── Service ──────────────────────────────────────────────────────────────
...
    AppointmentDetail:
      type: object
      properties:
        id:
          type: integer
        salonId:
          type: integer
        clientId:
          type: integer
        serviceId:
          type: integer
        employeeId:
          type: integer
        startsAt:
          type: string
        endsAt:
          type: string
        status:
          type: string
          enum: [scheduled, confirmed, completed, cancelled]
        notes:
          type: ["string", "null"]
        clientName:
          type: string
        clientWhatsapp:
          type: ["string", "null"]
        clientPhone:
          type: ["string", "null"]
        serviceName:
          type: string
        servicePrice:
          type: string
        employeeName:
          type: string
        createdAt:
          type: string
        paymentStatus:
          type: string
          enum: [not_due, pending, paid]
        paymentMethod:
          type: ["string", "null"]
          enum: [cash, pix, credit, debit, transfer, null]
        paymentAmount:
          type: ["string", "null"]
      required: [id, salonId, clientId, serviceId, employeeId, startsAt, endsAt, status, clientName, serviceName, servicePrice, employeeName, createdAt, paymentStatus]
    CreateAppointmentBody:
      type: object
      properties:
        clientId:
          type: integer
        serviceId:
          type: integer
        employeeId:
          type: integer
        startsAt:
          type: string
        notes:
          type: string
      required: [clientId, serviceId, employeeId, startsAt]
    UpdateAppointmentBody:
      type: object
      properties:
        status:
          type: string
          enum: [scheduled, confirmed, completed, cancelled]
        startsAt:
          type: string
        endsAt:
          type: string
        notes:
          type: string
        employeeId:
          type: integer
        serviceId:
          type: integer
        paymentStatus:
          type: string
          enum: [not_due, pending, paid]
        paymentMethod:
          type: string
          enum: [cash, pix, credit, debit, transfer]
        paymentAmount:
          type: string

    # ── Payment ──────────────────────────────────────────────────────────────
...